feat:更新部门管理功能

+++ src/main/java/com/crm/common/handler/FieldMetaObjectHandler.java
@@ -20,7 +20,7 @@ public class FieldMetaObjectHandler implements MetaObjectHandler {
         // 更新时间
         strictInsertFill(metaObject, Constant.UPDATE_TIME, LocalDateTime.class, LocalDateTime.now());
         // 删除标识
-        strictInsertFill(metaObject, Constant.DELETE_FLAG, Integer.class, 0);
+        strictInsertFill(metaObject, Constant.DELETE_FLAG, Byte.class, (byte) 0);
     }
 
     @Override
